Moved effects into Effects.hpp.
