+++ addons/map/Effects.hpp
@@ -1,0 +1,36 @@
+// "Smoke" effect classes for global flashlight glow
+
+class ACE_FlashlightEffect_White {
+    class Light1 {
+        simulation = "light";
+        type = "ACE_FlashlightLight_White";
+    };
+};
+
+class ACE_FlashlightEffect_Red {
+    class Light1 {
+        simulation = "light";
+        type = "ACE_FlashlightLight_Red";
+    };
+};
+
+class ACE_FlashlightEffect_Blue {
+    class Light1 {
+        simulation = "light";
+        type = "ACE_FlashlightLight_Blue";
+    };
+};
+
+class ACE_FlashlightEffect_Green {
+    class Light1 {
+        simulation = "light";
+        type = "ACE_FlashlightLight_Green";
+    };
+};
+
+class ACE_FlashlightEffect_Yellow {
+    class Light1 {
+        simulation = "light";
+        type = "ACE_FlashlightLight_Yellow";
+    };
+};
