+++ addons/map/config.cpp
@@ -30,6 +30,7 @@ class RscEdit;
 #include "CfgAmmo.hpp"
 #include "CfgSounds.hpp"
 #include "CfgLights.hpp"
+#include "Effects.hpp"
 
 class RscMapControl {
     maxSatelliteAlpha = 0.5;
@@ -160,42 +161,5 @@ class RscDisplayServerGetReady: RscDisplayGetReady {
         class TopRight: RscControlsGroup {
             #include "MapControls.hpp"
         };
-    };
-};
-
-// "Smoke" effect classes for global flashlight glow
-
-class ACE_FlashlightEffect_White {
-    class Light1 {
-        simulation = "light";
-        type = "ACE_FlashlightLight_White";
-    };
-};
-
-class ACE_FlashlightEffect_Red {
-    class Light1 {
-        simulation = "light";
-        type = "ACE_FlashlightLight_Red";
-    };
-};
-
-class ACE_FlashlightEffect_Blue {
-    class Light1 {
-        simulation = "light";
-        type = "ACE_FlashlightLight_Blue";
-    };
-};
-
-class ACE_FlashlightEffect_Green {
-    class Light1 {
-        simulation = "light";
-        type = "ACE_FlashlightLight_Green";
-    };
-};
-
-class ACE_FlashlightEffect_Yellow {
-    class Light1 {
-        simulation = "light";
-        type = "ACE_FlashlightLight_Yellow";
     };
 };
